feat: add async versions of the encrypt and decrypt functions
