+++ .changeset/dry-walls-care.md
@@ -1,0 +1,5 @@
+---
+"@sigmacomputing/node-embed-sdk": minor
+---
+
+Add async versions of encryption and decryption function
