+++ packages/node-embed-sdk/src/index.ts
@@ -1,1 +1,2 @@
 export { encrypt, decrypt } from "./encryption";
+export { encrypt as encryptAsync, decrypt as decryptAsync } from "./promises";
