+++ packages/node-embed-sdk/tsup.config.js
@@ -3,6 +3,7 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     index: "src/index.ts",
+    promises: "src/promises.ts",
   },
   sourcemap: false,
   minify: false,
