+++ packages/test-pack/README.md
@@ -1,0 +1,3 @@
+# Sigma Node.js Embed SDK
+
+This package tests the various entry points for the node-embed-sdk package.
