+++ packages/test-pack/tsconfig.lint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@sigmacomputing/typescript-config/base.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
